refactor: move special keys functions into another class
